added option to stop copying empty directories

+++ src/lib.rs
@@ -12,10 +12,11 @@ pub mod fil{
     - rmb to make a list of newly added contents
     */
 
-    pub fn copy_over(src: &Path, target: &Path, tgt_ext: &Vec<&'static str>, added_items: &mut Vec<PathBuf>) 
-    -> Result<(),&'static str>{
+    pub fn copy_over(src: &Path, target: &Path, tgt_ext: &Vec<&'static str>, 
+    added_items: &mut Vec<PathBuf>, is_copy_empty: bool) -> Result<(), String>{
+
         if target.starts_with(src){
-            return Err("Target directory within source directory");
+            return Err("Target directory within source directory".to_string());
         }
 
         let src_iter = get_dir_iter(src)?;
@@ -28,13 +29,14 @@ pub mod fil{
 
         // check which files do not exist, create new file and copy
         'outer: for src_file in src_iter.map(|x| x.unwrap()){
-            let Ok(src_type) = src_file.file_type() else { return Err("cannot get file type") };
+            let Ok(src_type) = src_file.file_type() else { return Err("Cannot get file type".to_string()) };
             match src_file.path().extension() {
                 Some(extension) => {
-                    if !tgt_ext.contains( &extension.to_str().expect("cannot get ext") ) { continue 'outer }
+                    // if not in tgt_ext and not a dir, ignore the file
+                    if !tgt_ext.contains( &extension.to_str().expect("Cannot get ext") ) { continue 'outer }
                 },
                 None => {
-                    // if not in tgt_ext or a directory, ignore the file
+                    // if no extension and not a directory, ignore the file
                     if !src_type.is_dir() { continue 'outer }
                 }
             }
@@ -42,9 +44,10 @@ pub mod fil{
             // check if there are matching names
             for file_in_target in &files_in_target{
                 if file_in_target.file_name().unwrap() == src_file.file_name(){
-                    // if is same directory repeat the copying in the directory
+                    // if is same directory, repeat the copying in the directory
+                    // otherwise just ignore same name files
                     if src_type.is_dir(){
-                        copy_over(&src_file.path(), &file_in_target, tgt_ext, added_items)?;
+                        copy_over(&src_file.path(), &file_in_target, tgt_ext, added_items, is_copy_empty)?;
                     }
                     continue 'outer;
                 }
@@ -54,7 +57,7 @@ pub mod fil{
             // use target directory and append name of src file to path
             if src_type.is_dir(){
                 create_and_copy_dir(&src_file.path(), &target.join(src_file.file_name()),
-                tgt_ext, added_items)?;
+                tgt_ext, added_items, is_copy_empty)?;
             }
             else{
                 create_and_copy_file(&src_file.path(), &target.join(src_file.file_name()), 
@@ -75,30 +78,40 @@ pub mod fil{
 
     fn create_and_copy_file(src: &Path, target: &Path, added_items: &mut Vec<PathBuf>) -> Result<(), &'static str> {
         // check not same path
-        if src == target { return Err("same path given, cannot copy file") }
+        if src == target { return Err("Same path given, cannot copy file") }
 
         // create and log created file
-        if let Err(_) = std::fs::File::create(target){return Err("cannot create new file")}
+        if let Err(_) = std::fs::File::create(target){return Err("Cannot create new file")}
         log(&format!(">> Created file: {}", target.display()));
         added_items.push(target.to_path_buf());
 
         // copy file
-        if let Err(_) =  std::fs::copy(src, target){return Err("cannot copy content")}
+        if let Err(_) =  std::fs::copy(src, target){return Err("Cannot copy content")}
 
         Ok(())
     }
 
-    fn create_and_copy_dir(src: &Path, target: &Path, tgt_ext: &Vec<&'static str>, added_items: &mut Vec<PathBuf>) 
-    -> Result<(), &'static str> {
+    // Result<bool> represents if the directory created is empty
+    fn create_and_copy_dir(src: &Path, target: &Path, tgt_ext: &Vec<&'static str>,
+    added_items: &mut Vec<PathBuf>, is_copy_empty: bool) -> Result<bool, String> {
         // check not same path
-        if src == target { return Err("same path given, cannot copy folder") }
+        if src == target { return Err("Same path given, cannot copy folder".to_string()) }
 
         // create new dir and log, then get iter
-        if let Err(_) = std::fs::create_dir(target){ return Err("cannot create new file") }
+        if let Err(_) = std::fs::create_dir(target){ return Err("Cannot create new file".to_string()) }
         log(&format!(">> Created directory: {}", target.display()));
         added_items.push(target.to_path_buf());
 
-        let Ok(src_dir) = get_dir_iter(src) else { return Err("cannot get copying dir") };
+        let Ok(src_dir) = get_dir_iter(src) else { 
+            return Err( format!("Cannot get source directory: {}", src.display()) );
+        };
+
+        // is_delete true by default
+        // makes file -> dont delete, set false
+        // makes dir -> dont delete, but dir can get deleted
+        // pass back if is deleted, reuse the result value and pass bool
+
+        let mut is_empty_dir = true;
 
         for file in src_dir.map(|x| x.unwrap()){
             match file.path().extension() {
@@ -110,43 +123,67 @@ pub mod fil{
                     // target extension then create and copy file
                     create_and_copy_file(&file.path(), &target.join(file.file_name()),
                     added_items)?;
+                    is_empty_dir = false;
                 },
                 None => {
                     // ignore if file is not a directory
                     if !file.file_type().unwrap().is_dir() { continue }
 
                     // create and copy dir
-                    create_and_copy_dir(&file.path(), &target.join(file.file_name()),
-                    tgt_ext, added_items)?;
+                    if !create_and_copy_dir(&file.path(), &target.join(file.file_name()),
+                    tgt_ext, added_items, is_copy_empty)?{
+                        is_empty_dir = false;
+                    }
                 }
             }
         }
-        Ok(())
+        
+        // abort if empty
+        if is_empty_dir && !is_copy_empty{
+            if let Err(_) = std::fs::remove_dir(target){
+                return Err( format!("Unable to delete supposed empty directory: {}", target.display()));
+            }
+            log(&format!(">> Deleted empty directory: {}", target.display()));
+            added_items.pop().expect("Cant find created dir?");
+
+            // return true since directory is gone
+            return Ok(true);
+        }
+
+        // return false since if directory was empty function would have returned true in previous line
+        Ok(false)
     }
 
     // directory is added, then files inside
     // reversing iter allows files to be deleted first before directory 
     // so path to inner files wont be invalidated if directory is deleted
-    pub fn delete_files<'a> (targets: &'a mut Vec<PathBuf>) -> Result<(), String> {
-        for file in targets.iter().rev(){
+    // iter uses target clone to allow deleted files to be popped out
+    // since going in reverse, the front of iter is the back of original targets
+    pub fn delete_files(targets: &mut Vec<PathBuf>) -> Result<(), String> {
+        for file in targets.clone().iter().rev(){
+            if !file.exists(){
+                return Err( format!("File or directory does not exist: {}", file.display()) );
+            }
+
             if file.is_dir() {
                 if let Err(_) = std::fs::remove_dir(file){
-                    return Err( format!("unable to delete directory: {}", 
-                    file.file_name().unwrap().to_str().unwrap()) );
+                    return Err( format!("Unable to delete directory: {}", 
+                    file.display()) );
                 }
+                targets.pop();
                 log(&format!(">> Deleted directory: {}", file.display()));
             }
             else{
                 if let Err(_) = std::fs::remove_file(file){
-                    return Err(format!("unable to delete file: {}", 
-                    file.file_name().unwrap().to_str().unwrap()));
+                    return Err(format!("Unable to delete file: {}", 
+                    file.display()) );
                 }
+                targets.pop();
                 log(&format!(">> Deleted file: {}", file.display()));
             }
         }
 
-        // remove all paths so they wont be pointing to deleted files
-        targets.clear();
+        // success
         Ok(())
     }
 }
